+++ src/components/MessagesList/MessagesList.module.scss
@@ -5,28 +5,59 @@
   --containerBorderColor: var(--borderColor, var(--textColor));
 
   height: 70%;
-  @include dFlex(column, flex-end);
+  @include dFlex(column, space-between);
   align-items: flex-start;
   overflow: visible;
+
+  .settingsCallerContainer {
+    width: 100%;
+    padding: 0.5rem;
+    @include dFlex(row, flex-end);
+
+    button {
+      --size: 3.5rem;
+
+      @include dFlex();
+      width: var(--size);
+      height: var(--size);
+      font-size: 2.5rem;
+      background: var(--secondaryAltColorDark);
+      color: var(--greyColorLight);
+      border-radius: 50%;
+
+      &:hover {
+        transform: rotate(200deg);
+        color: var(--greyColorLighter);
+      }
+    }
+  }
 
   .messagesContainer {
     border-bottom: none;
     width: 100%;
     overflow: auto;
-    & > pre {
+    @include dFlex(column);
+    align-items: flex-start;
+    gap: 0.5rem;
+    & > span {
       position: relative;
-      width: 100%;
+      margin-left: 0.5rem;
+      padding: 0.4rem 0.6rem;
+      min-width: 20%;
+      max-width: 70%;
+      border-radius: 5px;
 
-      & > p {
+      & > dl {
         position: relative;
         z-index: 2;
       }
 
       & > span {
         @include fullContainerAbsolute();
-        background: var(--bgLineColor);
-        opacity: 0.3;
+        background: var(--mainColor);
+        opacity: 0.35;
         z-index: 0;
+        border-radius: 5px;
       }
     }
   }
